backend/requirements.txt
black==22.10.0
Django==4.1.2
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
Pillow==9.2.0
